requirements.txt
et-xmlfile==1.1.0
numpy>=1.24.0
openpyxl>=3.1.0
pandas>=1.5.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.41
tqdm==4.64.0
boto3==1.21.39
requests==2.26.0
pyarrow>=8.0.0
fastparquet>=2023.1.0